refactor(index): make index and version available from journalist_app.api2.shared
